pinia decomposition
